Move the novel model to the models folder
