+++ tests/core/test_index.py
@@ -31,9 +31,9 @@ from novelwriter import SHARED
 from novelwriter.constants import nwFiles
 from novelwriter.core.index import Index, TagsIndex
 from novelwriter.core.item import ProjectItem
-from novelwriter.core.novelmodel import NovelModel
 from novelwriter.core.project import NWProject
 from novelwriter.enum import nwComment, nwItemClass, nwItemLayout, nwNovelExtra
+from novelwriter.models.novelmodel import NovelModel
 
 from tests.helpers import C, buildTestProject, cmpFiles
 from tests.mocked import causeException
